Bump RuntimeCompiler 5.0.0-alpha.1 + Marten 9.0.0-alpha.2 + revert Marten-side global:: qualifiers
Picks up the upstream packages that pair against the post-TPL-Dataflow-
removal JasperFx 2.0.0-alpha.13 + dedupe-pillar JasperFx.Events 2.0.0-
alpha.8:

- JasperFx.RuntimeCompiler  2.0.0-alpha.2 -> 5.0.0-alpha.1
- Marten                    9.0.0-alpha.1 -> 9.0.0-alpha.2
- Marten.AspNetCore         9.0.0-alpha.1 -> 9.0.0-alpha.2
- Marten.Newtonsoft         (new) 9.0.0-alpha.2

The Marten alpha.1 binary referenced the pre-removal 4-arg
RetryBlock<T>(..., Action<ExecutionDataflowBlockOptions>) ctor and was
binary-incompatible with the alpha.13 runtime — every DI resolution
that flowed into Marten.Events.EventGraph.Initialize raised
MissingMethodException. alpha.2 was rebuilt against alpha.13 and has
zero `System.Threading.Tasks.Dataflow` string references.

Marten alpha.2 dedupe-pillar adoption (jasperfx/marten#4403) drops
Marten.Events.Dcb.IEventBoundary<T>, Marten.Events.IEventStream<T>,
Marten.Events.IEventStoreOperations, and Marten.Events.StreamCompactingRequest<T>
in favor of the lifted JasperFx.Events counterparts. Reverted the
`global::Marten.Events.*` qualifiers I added in d215f72ec /
0f0139e59 (transitional disambiguation while both products shipped
duplicates) back to bare references.

Cascade fixes surfaced by the alpha.2 API:

- `using JasperFx.Events;` added to Wolverine.Marten source +
  Wolverine.Marten test code that imported `using Marten.Events;`
  alone (which used to resolve IEventStream<> to Marten's duplicate;
  now resolves to the non-generic Marten.Events.IEventStream).
- `using JasperFx.Events.Tags;` added to files referencing
  IEventBoundary<> after Marten.Events.Dcb.IEventBoundary<> was dropped.
- Wolverine.Saga.Version widened from int to long to match
  Marten.Metadata.IRevisioned.Version (widened to long in alpha.2);
  cascaded the same widening to Saga subclasses that explicitly opt
  into IRevisioned (UserRegistrationSaga, SideEffects1,
  GpStreamAggregate, GSimpleAggregate).
- WolverineProjectionCoordinator<T> generic constraint tightened from
  `where T : IDocumentStore` to `where T : class, IDocumentStore` —
  Marten.Events.Daemon.Coordination.IProjectionCoordinator<T> now
  carries a `where T : class` constraint.
- HostExtensions / TestingExtensions methods constrained on
  `where T : IDocumentStore` similarly tightened to `class, IDocumentStore`.
- `using Marten.Newtonsoft;` (+ Marten.Newtonsoft package pin) added
  to MartenTests/Bugs/Bug_1427 — Marten alpha.2 extracted the
  Newtonsoft serialization extension methods to a separate package.
- Reverted leftover ConcurrencyStyle qualifier in
  MartenTests/AggregateHandlerWorkflow/marten_command_workflow_middleware.cs
  to global::Wolverine.Marten side (matches the AggregateHandler
  attribute's enum type).
- Tightened RegisterBoundaryEventsFrame<T> /
  ApplyBoundaryEventsFromAsyncEnumerableFrame<T> /
  RegisterEventsFrame<T> from `where T : notnull` to `where T : class`
  to satisfy IEventBoundary<T>'s / IEventStream<T>'s `class` constraints.

Local verification (per CLAUDE.md): full `dotnet build wolverine.slnx
-c Release` is clean (0 warnings / 0 errors on net9.0 + net10.0).
CoreTests: 1690/1692 passing; the 2 failures
(`system_message_type_filtering.observer_does_not_receive_message_routed_for_system_types`
× 2 TFMs) are pre-existing on main (run 25883603362) and unrelated to
this pin bump.

+++ src/Persistence/MartenTests/end_to_end_publish_messages_through_marten_to_wolverine.cs
@@ -257,7 +257,7 @@ public class SideEffects1 : IRevisioned
     public int B { get; set; }
     public int C { get; set; }
     public int D { get; set; }
-    public int Version { get; set; }
+    public long Version { get; set; }
 }
 
 // Wrap it in your own IHostedService
@@ -434,7 +434,7 @@ public class Order2
 
     // This is important, by Marten convention this would
     // be the
-    public int Version { get; set; }
+    public long Version { get; set; }
 
     public Order2(OrderCreated created)
     {
