+++ src/Wolverine/Saga.cs
@@ -29,9 +29,13 @@ public abstract class Saga
     /// <summary>
     /// For saga providers that support this, this is a version of the saga to help enforce optimistic concurrency
     /// protections. This value is the current version that is stored by saga storage and will
-    /// be incremented upon save
+    /// be incremented upon save.
+    /// Widened to <see cref="long"/> in 6.0 to align with Marten's
+    /// <c>Marten.Metadata.IRevisioned.Version</c> (which is <see cref="long"/>
+    /// from Marten 9.0.0-alpha.2 onward), so sagas can implement <c>IRevisioned</c>
+    /// without a shadow override.
     /// </summary>
-    public int Version { get; set; }
+    public long Version { get; set; }
 }
 
 #endregion
